Docs on core resources
