+++ cognite/src/api/core/assets.rs
@@ -1,8 +1,7 @@
 use crate::api::resource::*;
 use crate::dto::core::asset::*;
-use crate::dto::items::ItemsWithCursor;
 use crate::error::Result;
-use crate::{Identity, Patch};
+use crate::{Identity, ItemsWithoutCursor, Patch};
 
 /// Assets represent objects or groups of objects from the physical world.
 /// Assets are organized in hierarchies. For example, a water pump asset can
@@ -25,17 +24,41 @@ impl AssetsResource {
     /// Retrieve a list of assets by their IDs.
     ///
     /// Will fail if `ignore_unknown_ids` is false and the assets are not present in CDF.
+    ///
+    /// # Arguments
+    ///
+    /// * `asset_ids` - List of IDs or external IDs to retrieve.
+    /// * `ignore_unknown_ids` - If `true`, missing assets will be ignored, instead of causing
+    /// the request to fail.
+    /// * `aggregated_properties` - List of aggregated properties to include in response.
     pub async fn retrieve(
         &self,
         asset_ids: &[Identity],
         ignore_unknown_ids: bool,
-        aggregated_properties: Option<Vec<String>>,
+        aggregated_properties: Option<Vec<AssetAggregatedProperty>>,
     ) -> Result<Vec<Asset>> {
         let mut id_items = RetrieveAssetsRequest::from(asset_ids);
         id_items.ignore_unknown_ids = ignore_unknown_ids;
         id_items.aggregated_properties = aggregated_properties;
-        let assets_response: ItemsWithCursor<Asset> =
+        let assets_response: ItemsWithoutCursor<Asset> =
             self.api_client.post("assets/byids", &id_items).await?;
         Ok(assets_response.items)
+    }
+
+    /// Compute aggregates over assets, such as getting the count of all assets in a project,
+    /// checking different names and descriptions of assets in your project, etc.
+    ///
+    /// # Arguments
+    ///
+    /// * `aggregate` - Aggregate to compute
+    ///
+    /// The returned aggregates depend on which aggregates were requested.
+    pub async fn aggregate(
+        &self,
+        aggregate: AssetAggregateRequest,
+    ) -> Result<Vec<AssetAggregateResponse>> {
+        let resp: ItemsWithoutCursor<AssetAggregateResponse> =
+            self.api_client.post("assets/aggregate", &aggregate).await?;
+        Ok(resp.items)
     }
 }
