+++ cognite/src/api/data_ingestion/raw.rs
@@ -11,6 +11,11 @@ pub type RawResource = Resource<RawRow>;
 
 impl RawResource {
     /// List Raw databases in the project.
+    ///
+    /// # Arguments
+    ///
+    /// * `limit` - Maximum number of databases to retrieve.
+    /// * `cursor` - Optional cursor for pagination.
     pub async fn list_databases(
         &self,
         limit: Option<i32>,
@@ -23,6 +28,10 @@ impl RawResource {
     }
 
     /// Create a list of Raw databases.
+    ///
+    /// # Arguments
+    ///
+    /// * `dbs` - Databases to create.
     pub async fn create_databases(&self, dbs: &[Database]) -> Result<Vec<Database>> {
         let items = Items::from(dbs);
         let result: ItemsWithCursor<Database> = self.api_client.post("raw/dbs", &items).await?;
@@ -30,6 +39,10 @@ impl RawResource {
     }
 
     /// Delete a list of raw databases.
+    ///
+    /// # Arguments
+    ///
+    /// * `to_delete` - Request describing which databases to delete and how.
     pub async fn delete_databases(&self, to_delete: &DeleteDatabasesRequest) -> Result<()> {
         self.api_client
             .post::<::serde_json::Value, DeleteDatabasesRequest>("raw/dbs/delete", to_delete)
@@ -38,6 +51,12 @@ impl RawResource {
     }
 
     /// List tables in a a raw database.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to list tables in.
+    /// * `limit` - Maximum number of tables to retrieve.
+    /// * `cursor` - Optional cursor for pagination.
     pub async fn list_tables(
         &self,
         db_name: &str,
@@ -50,13 +69,21 @@ impl RawResource {
     }
 
     /// Create tables in a raw database.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to create tables in.
+    /// * `ensure_parent` - If this is set to `true`, create database if it doesn't already exist.
+    /// * `tables` - Tables to create.
     pub async fn create_tables(
         &self,
         db_name: &str,
-        ensure_parent: Option<bool>,
+        ensure_parent: bool,
         tables: &[Table],
     ) -> Result<Vec<Table>> {
-        let query = EnsureParentQuery { ensure_parent };
+        let query = EnsureParentQuery {
+            ensure_parent: Some(ensure_parent),
+        };
         let path = format!("raw/dbs/{db_name}/tables");
         let items = Items::from(tables);
         let result: ItemsWithCursor<Table> = self
@@ -67,6 +94,11 @@ impl RawResource {
     }
 
     /// Delete tables in a raw database.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to delete tables from.
+    /// * `to_delete` - Tables to delete.
     pub async fn delete_tables(&self, db_name: &str, to_delete: &[Table]) -> Result<()> {
         let path = format!("raw/dbs/{db_name}/tables/delete");
         let items = Items::from(to_delete);
@@ -78,6 +110,12 @@ impl RawResource {
 
     /// Retrieve cursors for parallel reads. This can be used to efficiently download
     /// large volumes of data from a raw table in parallel.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to retrieve from.
+    /// * `table_name` - Table to retrieve from.
+    /// * `params` - Optional filter parameters.
     pub async fn retrieve_cursors_for_parallel_reads(
         &self,
         db_name: &str,
@@ -91,6 +129,12 @@ impl RawResource {
     }
 
     /// Retrieve rows from a table, with some basic filtering options.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to retrieve rows from.
+    /// * `table_name` - Table to retrieve rows from.
+    /// * `params` - Optional filter parameters.
     pub async fn retrieve_rows(
         &self,
         db_name: &str,
@@ -104,6 +148,13 @@ impl RawResource {
     /// Insert rows into a table.
     ///
     /// If `ensure_parent` is true, create the database and/or table if they do not exist.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to insert rows into.
+    /// * `table_name` - Table to insert rows into.
+    /// * `ensure_parent` - Create database and/or table if they do not exist.
+    /// * `rows` - Raw rows to create.
     pub async fn insert_rows(
         &self,
         db_name: &str,
@@ -125,12 +176,24 @@ impl RawResource {
     }
 
     /// Retrieve a single row from a raw table.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to retrieve from.
+    /// * `table_name` - Table to retrieve from.
+    /// * `key` - Key of row to retrieve.
     pub async fn retrieve_row(&self, db_name: &str, table_name: &str, key: &str) -> Result<RawRow> {
         let path = format!("raw/dbs/{db_name}/tables/{table_name}/rows/{key}");
         self.api_client.get(&path).await
     }
 
     /// Delete rows from a raw table.
+    ///
+    /// # Arguments
+    ///
+    /// * `db_name` - Database to delete from.
+    /// * `table_name` - Table to delete from.
+    /// * `to_delete` - Rows to delete.
     pub async fn delete_rows(
         &self,
         db_name: &str,
