+++ cognite/src/api/data_organization/datasets.rs
@@ -3,12 +3,7 @@ use crate::dto::{data_organization::datasets::*, items::ItemsWithCursor};
 use crate::error::Result;
 use crate::{Filter, Identity, Patch};
 
-/// Data sets let you document and track data lineage, as well as
-/// restrict access to data.
-///
-/// Data sets group and track data by its source.
-/// For example, a data set can contain all work orders originating from SAP.
-/// Typically, an organization will have one data set for each of its data ingestion pipelines in CDF.
+/// API resource for data sets.
 pub type DataSetsResource = Resource<DataSet>;
 
 impl WithBasePath for DataSetsResource {
@@ -21,6 +16,11 @@ impl Update<Patch<PatchDataSet>, DataSet> for DataSetsResource {}
 impl FilterItems<DataSetFilter, DataSet> for DataSetsResource {}
 
 impl DataSetsResource {
+    /// Calculate the total number of data sets in the project matching the given filter
+    ///
+    /// # Arguments
+    ///
+    /// * `filter` - Optional filter.
     pub async fn count(&self, filter: DataSetFilter) -> Result<DataSetsCount> {
         let query = Filter::<DataSetFilter>::new(filter, None, None);
         let result: ItemsWithCursor<DataSetsCount> =
