+++ cognite/src/dto/core/asset/mod.rs
@@ -1,5 +1,7 @@
+mod aggregate;
 mod filter;
 
+pub use self::aggregate::*;
 pub use self::filter::*;
 
 use crate::dto::identity::Identity;
@@ -15,9 +17,13 @@ use super::GeoLocation;
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Aggregated asset properties.
 pub struct AssetAggregate {
+    /// Number of direct descendants of the asset.
     pub child_count: Option<i32>,
+    /// Asset path depth (number of levels below root node)
     pub depth: Option<i32>,
+    /// IDs of assets on the path to the asset.
     pub path: Option<Vec<CogniteId>>,
 }
 
@@ -25,24 +31,51 @@ pub struct AssetAggregate {
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct Asset {
+    /// Server-generated ID of the asset.
     pub id: i64,
+    /// Name of the asset.
     pub name: String,
+    /// Unique user-provided external ID.
     pub external_id: Option<String>,
+    /// ID of the parent node.
     pub parent_id: Option<i64>,
+    /// External ID of the parent node.
     pub parent_external_id: Option<String>,
+    /// Description of the asset.
     pub description: Option<String>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes,
+    /// value 10240 bytes, up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<HashMap<String, String>>,
+    /// Source of the asset.
     pub source: Option<String>,
+    /// Time this asset was created, in milliseconds since epoch.
     pub created_time: i64,
+    /// Time this assset was last updated, in milliseconds since epoch.
     pub last_updated_time: i64,
+    /// ID of the root asset.
     pub root_id: Option<i64>,
+    /// Aggregated metrics computed on this asset.
     pub aggregates: Option<AssetAggregate>,
+    /// ID of the data set this asset belongs to.
     pub data_set_id: Option<i64>,
+    /// List of the labels associated with this asset.
     pub labels: Option<Vec<CogniteExternalId>>,
+    /// Geographic metadata.
     pub geo_location: Option<GeoLocation>,
 }
 
 impl Asset {
+    /// Create an asset
+    ///
+    /// # Arguments
+    ///
+    /// * `name` - Name of the asset.
+    /// * `description` - Description of the asset.
+    /// * `external_id` - External ID of the asset, must be unique.
+    /// * `parent_id` - ID of the parent of this asset.
+    /// * `metadata` - Optional application specific metadata.
+    /// * `source` - Source of this asset.
     pub fn new(
         name: &str,
         description: &str,
@@ -74,16 +107,37 @@ impl Asset {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Create an asset.
 pub struct AddAsset {
+    /// Name of the asset.
     pub name: String,
+    /// Unique external ID of the asset.
     pub external_id: Option<String>,
+    /// Parent node ID used to specify a parent-child relationship. Do not use
+    /// in combination with `parent_external_id`
     pub parent_id: Option<i64>,
+    /// Description of the asset.
     pub description: Option<String>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes, value 10240 bytes,
+    /// up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<HashMap<String, String>>,
+    /// Source of the asset.
     pub source: Option<String>,
+    /// External ID of the parent asset.
+    ///  When specifying this field, the API will resolve the external ID
+    /// into an internal ID and use the internal ID to store the
+    /// parent-child relation. As a result, a later change to update the parent's
+    /// external ID will not affect this parent-child relationship as
+    /// it is based on internal ID.
+    ///
+    /// Do not use in combination with `parent_id`
     pub parent_external_id: Option<String>,
+    /// ID of the dataset this asset belongs to.
     pub data_set_id: Option<i64>,
+    /// A list of labels associated with this asset.
     pub labels: Option<Vec<CogniteExternalId>>,
+    /// Geographic metadata.
     pub geo_location: Option<GeoLocation>,
 }
 
@@ -120,16 +174,37 @@ impl EqIdentity for AddAsset {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Update an asset.
 pub struct PatchAsset {
+    /// Unique external ID of the asset.
     pub external_id: Option<UpdateSetNull<String>>,
+    /// Name of the asset.
     pub name: Option<UpdateSet<String>>,
+    /// Description of the asset.
     pub description: Option<UpdateSetNull<String>>,
+    /// ID of the dataset this asset belongs to.
     pub data_set_id: Option<UpdateSetNull<i64>>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes, value 10240 bytes,
+    /// up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<UpdateMap<String, String>>,
+    /// Source of the asset.
     pub source: Option<UpdateSetNull<String>>,
+    /// Parent node ID used to specify a parent-child relationship. Do not use
+    /// in combination with `parent_external_id`
     pub parent_id: Option<UpdateSet<i64>>,
+    /// External ID of the parent asset.
+    ///  When specifying this field, the API will resolve the external ID
+    /// into an internal ID and use the internal ID to store the
+    /// parent-child relation. As a result, a later change to update the parent's
+    /// external ID will not affect this parent-child relationship as
+    /// it is based on internal ID.
+    ///
+    /// Do not use in combination with `parent_id`
     pub parent_external_id: Option<UpdateSet<String>>,
+    /// A list of labels associated with this asset.
     pub labels: Option<UpdateList<CogniteExternalId, CogniteExternalId>>,
+    /// Geographic metadata.
     pub geo_location: Option<UpdateSetNull<GeoLocation>>,
 }
 
@@ -186,10 +261,10 @@ impl From<Asset> for Patch<PatchAsset> {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
-pub struct RetrieveAssetsRequest {
+pub(crate) struct RetrieveAssetsRequest {
     pub items: ::serde_json::Value,
     pub ignore_unknown_ids: bool,
-    pub aggregated_properties: Option<Vec<String>>,
+    pub aggregated_properties: Option<Vec<AssetAggregatedProperty>>,
 }
 
 impl<T: Serialize> From<&Vec<T>> for RetrieveAssetsRequest {
@@ -214,8 +289,12 @@ impl<T: Serialize> From<&[T]> for RetrieveAssetsRequest {
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Request for deleting a list of assets.
 pub struct DeleteAssetsRequest {
+    /// Internal or external IDs of assets to delete.
     pub items: Vec<Identity>,
+    /// If `true`, ignore any IDs that do not exist in CDF.
     pub ignore_unknown_ids: bool,
+    /// If `true`, recursively delete any children of the deleted assets.
     pub recursive: bool,
 }
