+++ cognite/src/dto/core/files/file.rs
@@ -7,52 +7,82 @@ use crate::{
     UpdateMap, UpdateSetNull,
 };
 
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct FileListResponse {
-    pub items: Vec<FileMetadata>,
-    next_cursor: Option<String>,
-}
-
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Description of a CDF file.
 pub struct FileMetadata {
+    /// File external ID. Must be unique accross all files in the project.
     pub external_id: Option<String>,
+    /// File name.
     pub name: String,
+    /// Directory containing the file. Must be an absolute, unix-style path.
     pub directory: Option<String>,
+    /// Source of the file.
     pub source: Option<String>,
+    /// File mime type, e.g. `application/pdf`.
     pub mime_type: Option<String>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes, value 10240 bytes,
+    /// up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<HashMap<String, String>>,
+    /// List of assets the file is tied to.
     pub asset_ids: Option<Vec<i64>>,
+    /// Data set the file belongs to.
     pub data_set_id: Option<i64>,
+    /// Timestamp in milliseconds since epoch when this file was created in the source system.
     pub source_created_time: Option<i64>,
+    /// Timestamp in milliseconds since epoch when this file was last modified in the source system.
     pub source_modified_time: Option<i64>,
+    /// The required security categories to access this file.
     pub security_categories: Option<Vec<i64>>,
+    /// List of labels associated with this file.
     pub labels: Option<Vec<CogniteExternalId>>,
+    /// File internal ID.
     pub id: i64,
+    /// Whether or not the actual file is uploaded.
     pub uploaded: bool,
+    /// Time this file was uploaded, in milliseconds since epoch.
     pub uploaded_time: Option<i64>,
+    /// Time this file was created, in milliseconds since epoch.
     pub created_time: i64,
+    /// Time this file was last modified, in milliseconds since epoch.
     pub last_updated_time: i64,
+    /// URL for uploading data to this file. Returned only in response to
+    /// `upload`.
     pub upload_url: Option<String>,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Create a new file.
 pub struct AddFile {
+    /// File external ID. Must be unique accross all files in the project.
     pub external_id: Option<String>,
+    /// File name.
     pub name: String,
+    /// Directory containing the file. Must be an absolute, unix-style path.
     pub directory: Option<String>,
+    /// Source of the file.
     pub source: Option<String>,
+    /// File mime type, e.g. `application/pdf`.
     pub mime_type: Option<String>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes, value 10240 bytes,
+    /// up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<HashMap<String, String>>,
+    /// List of assets the file is tied to.
     pub asset_ids: Option<Vec<i64>>,
+    /// Data set the file belongs to.
     pub data_set_id: Option<i64>,
+    /// Timestamp in milliseconds since epoch when this file was created in the source system.
     pub source_created_time: Option<i64>,
+    /// Timestamp in milliseconds since epoch when this file was last modified in the source system.
     pub source_modified_time: Option<i64>,
+    /// The required security categories to access this file.
     pub security_categories: Option<Vec<i64>>,
+    /// List of labels associated with this file.
     pub labels: Option<Vec<CogniteExternalId>>,
 }
 
@@ -87,17 +117,31 @@ impl EqIdentity for AddFile {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Update a file.
 pub struct PatchFile {
+    /// File external ID. Must be unique accross all files in the project.
     pub external_id: Option<UpdateSetNull<String>>,
+    /// Directory containing the file. Must be an absolute, unix-style path.
     pub directory: Option<UpdateSetNull<String>>,
+    /// Source of the file.
     pub source: Option<UpdateSetNull<String>>,
+    /// File mime type, e.g. `application/pdf`.
     pub mime_type: Option<UpdateSetNull<String>>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Limits: Maximum length of key is 128 bytes, value 10240 bytes,
+    /// up to 256 key-value pairs, of total size at most 10240.
     pub metadata: Option<UpdateMap<String, String>>,
+    /// List of assets the file is tied to.
     pub asset_ids: Option<UpdateList<i64, i64>>,
+    /// Timestamp in milliseconds since epoch when this file was created in the source system.
     pub source_created_time: Option<UpdateSetNull<i64>>,
+    /// Timestamp in milliseconds since epoch when this file was last modified in the source system.
     pub source_modified_time: Option<UpdateSetNull<i64>>,
+    /// Data set the file belongs to.
     pub data_set_id: Option<UpdateSetNull<i64>>,
+    /// The required security categories to access this file.
     pub security_categories: Option<UpdateList<i64, i64>>,
+    /// List of labels associated with this file.
     pub labels: Option<UpdateList<CogniteExternalId, CogniteExternalId>>,
 }
 
@@ -148,14 +192,19 @@ impl From<FileMetadata> for Patch<PatchFile> {
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Download URL for a file
 pub struct FileDownloadUrl {
     #[serde(flatten)]
+    /// ID of the file.
     pub id: Identity,
+    /// Temporary download URL for the file.
     pub download_url: String,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Aggregates on files.
 pub struct FileAggregates {
+    /// Number of files in the project.
     pub count: i64,
 }
