+++ cognite/src/dto/core/sequences.rs
@@ -1,5 +1,5 @@
 use crate::{
-    models::FdmFilter, EqIdentity, Identity, IntoPatch, IntoPatchItem, Partition, Patch, Range,
+    AdvancedFilter, EqIdentity, Identity, IntoPatch, IntoPatchItem, Partition, Patch, Range,
     SetCursor, UpdateList, UpdateMap, UpdateSet, UpdateSetNull, WithPartition,
 };
 
@@ -11,28 +11,42 @@ use super::CoreSortItem;
 
 #[derive(Serialize, Deserialize, Debug, Clone, Copy, Default)]
 #[serde(rename_all = "camelCase")]
+/// Type of sequence value.
 pub enum SequenceValueType {
     #[serde(rename = "DOUBLE")]
     #[default]
+    /// Double precision floating point.
     Double,
     #[serde(rename = "STRING")]
+    /// String.
     String,
     #[serde(rename = "LONG")]
+    /// 64-bit integer.
     Long,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
 #[serde(rename_all = "camelCase")]
+/// Description of a single sequence column
 pub struct SequenceColumn {
+    /// Name of the sequence column.
     pub name: Option<String>,
+    /// External ID of the sequence column. Must be unique for a given sequence.
     pub external_id: String,
+    /// Description of the sequence column.
     pub description: Option<String>,
+    /// Type of column value.
     pub value_type: SequenceValueType,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Maximum length of key is 128 bytes, up to 256 key-value pairs,
+    /// up to a total size of 10000 bytes across all keys and values.
     pub metadata: Option<HashMap<String, String>>,
     #[serde(skip_serializing)]
+    /// Time this sequence column was created, in milliseconds since epoch.
     pub created_time: Option<i64>,
     #[serde(skip_serializing)]
+    /// Time this sequence column was last updated, in milliseconds since epoch.
     pub last_updated_time: Option<i64>,
 }
 
@@ -40,15 +54,27 @@ pub struct SequenceColumn {
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct Sequence {
+    /// Sequence internal ID.
     pub id: i64,
+    /// External ID of the sequence. Must be unique within the project.
     pub external_id: Option<String>,
+    /// Name of the sequence.
     pub name: Option<String>,
+    /// Description of the sequence.
     pub description: Option<String>,
+    /// ID of asset this sequence belongs to.
     pub asset_id: Option<i64>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Maximum length of key is 128 bytes, up to 256 key-value pairs,
+    /// up to a total size of 10000 bytes across all keys and values.
     pub metadata: Option<HashMap<String, String>>,
+    /// List of columns in this sequence.
     pub columns: Vec<SequenceColumn>,
+    /// Time this sequence was created, in milliseconds since epoch.
     pub created_time: i64,
+    /// Time this sequence was last updated, in milliseconds since epoch.
     pub last_updated_time: i64,
+    /// ID of the data set this sequence belongs to.
     pub data_set_id: Option<i64>,
 }
 
@@ -56,12 +82,21 @@ pub struct Sequence {
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct AddSequence {
+    /// External ID of the sequence. Must be unique within the project.
     pub external_id: Option<String>,
+    /// Name of the sequence.
     pub name: Option<String>,
+    /// Description of the sequence.
     pub description: Option<String>,
+    /// ID of asset this sequence belongs to.
     pub asset_id: Option<i64>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Maximum length of key is 128 bytes, up to 256 key-value pairs,
+    /// up to a total size of 10000 bytes across all keys and values.
     pub metadata: Option<HashMap<String, String>>,
+    /// List of columns in this sequence.
     pub columns: Vec<SequenceColumn>,
+    /// ID of the data set this sequence belongs to.
     pub data_set_id: Option<i64>,
 }
 
@@ -91,9 +126,13 @@ impl EqIdentity for AddSequence {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Update the columns of a sequence.
 pub struct UpdateSequenceColumns {
+    /// List of column updates.
     pub modify: Option<Vec<Patch<PatchSequenceColumn>>>,
+    /// List of column definitions to add.
     pub add: Option<Vec<SequenceColumn>>,
+    /// List of columns to remove.
     pub remove: Option<Vec<String>>,
 }
 
@@ -117,10 +156,17 @@ impl From<UpdateList<SequenceColumn, String>> for UpdateSequenceColumns {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Update a sequence column.
 pub struct PatchSequenceColumn {
+    /// Description of the sequence.
     pub description: Option<UpdateSetNull<String>>,
+    /// External ID of the sequence column. Must be unique for a given sequence.
     pub external_id: Option<UpdateSet<String>>,
+    /// Name of the sequence column.
     pub name: Option<UpdateSetNull<String>>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Maximum length of key is 128 bytes, up to 256 key-value pairs,
+    /// up to a total size of 10000 bytes across all keys and values.
     pub metadata: Option<UpdateMap<String, String>>,
 }
 
@@ -128,12 +174,21 @@ pub struct PatchSequenceColumn {
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct PatchSequence {
+    /// Name of the sequence.
     pub name: Option<UpdateSetNull<String>>,
+    /// Description of the sequence.
     pub description: Option<UpdateSetNull<String>>,
+    /// ID of asset this sequence belongs to.
     pub asset_id: Option<UpdateSetNull<i64>>,
+    /// External ID of the sequence. Must be unique within the project.
     pub external_id: Option<UpdateSetNull<String>>,
+    /// Custom, application specific metadata. String key -> String value.
+    /// Maximum length of key is 128 bytes, up to 256 key-value pairs,
+    /// up to a total size of 10000 bytes across all keys and values.
     pub metadata: Option<UpdateMap<String, String>>,
+    /// ID of the data set this sequence belongs to.
     pub data_set_id: Option<UpdateSetNull<i64>>,
+    /// List of columns in this sequence.
     pub columns: Option<UpdateSequenceColumns>,
 }
 
@@ -177,27 +232,46 @@ impl From<Sequence> for Patch<PatchSequence> {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Filter on sequences.
 pub struct SequenceFilter {
+    /// Include sequences with this name.
     pub name: Option<String>,
+    /// Filter by this (case-sensitive) prefix for the external ID.
     pub external_id_prefix: Option<String>,
+    /// Filter by sequence metadata.
     pub metadata: Option<HashMap<String, String>>,
+    /// Include sequences associated with one of these assets.
     pub asset_ids: Option<Vec<i64>>,
+    /// Include sequences associated with an asset in the tree of one of these
+    /// root assets.
     pub root_asset_ids: Option<Vec<i64>>,
+    /// Include sequences associated with an asset in the subtree of one of these
+    /// assets.
     pub asset_subtree_ids: Option<Vec<Identity>>,
+    /// Range of timestamps for `created_time`.
     pub created_time: Option<Range<i64>>,
+    /// Range of timestamps for `last_updated_time`.
     pub last_updated_time: Option<Range<i64>>,
+    /// Include sequences which are tied to one of these data sets.
     pub data_set_ids: Option<Vec<Identity>>,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Request for filtering sequences.
 pub struct SequenceFilterRequest {
+    /// Simple sequences filter.
     pub filter: Option<SequenceFilter>,
-    pub advanced_filter: Option<FdmFilter>,
+    /// Advanced filter.
+    pub advanced_filter: Option<AdvancedFilter>,
+    /// Maximum number of sequences to return.
     pub limit: Option<i32>,
+    /// Optional cursor for pagination.
     pub cursor: Option<String>,
+    /// Split the result set into partitions.
     pub partition: Option<Partition>,
+    /// Sort the result by these properties. The order is significant.
     pub sort: Option<Vec<CoreSortItem>>,
 }
 
@@ -218,19 +292,29 @@ impl WithPartition for SequenceFilterRequest {
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Fuzzy search sequences.
 pub struct SequenceSearch {
+    /// Fuzzy search on name.
     pub name: Option<String>,
+    /// Fuzzy search on description.
     pub description: Option<String>,
+    /// Searches on name and description using wildcard search on each of the words
+    /// (separated by spaces). Retrieves results where at least one word must match.
     pub query: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
 #[serde(untagged)]
+/// Value of a sequence row.
 pub enum SequenceRowValue {
+    /// String value.
     String(String),
+    /// 64-bit integer value.
     Long(i64),
+    /// Double precision floating point value.
     Double(f64),
+    /// Null value.
     Null(()),
 }
 
@@ -242,67 +326,98 @@ impl Default for SequenceRowValue {
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// A single sequence row.
 pub struct SequenceRow {
+    /// Row number.
     pub row_number: i64,
+    /// List of values in order of requested columns.
     pub values: Vec<SequenceRowValue>,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// A simple reference to a sequence column.
 pub struct BasicSequenceColumn {
+    /// Sequence column external ID.
     pub external_id: String,
+    /// Sequence column name.
     pub name: Option<String>,
+    /// Column value type.
     pub value_type: SequenceValueType,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Response when retrieving sequence rows.
 pub struct RetrieveSequenceRowsResponse {
+    /// Internal ID of sequence.
     pub id: i64,
+    /// External ID of sequence.
     pub external_id: Option<String>,
+    /// List of requested columns in sequence.
     pub columns: Vec<BasicSequenceColumn>,
+    /// List of retrieved rows.
     pub rows: Vec<SequenceRow>,
+    /// Cursor for pagination.
     pub next_cursor: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Request to insert a list of rows into a sequence.
 pub struct InsertSequenceRows {
+    /// Column external IDs to insert into.
     pub columns: Vec<String>,
+    /// List of rows to insert.
     pub rows: Vec<SequenceRow>,
     #[serde(flatten)]
+    /// ID or external ID of sequence to insert into.
     pub id: Identity,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, Default, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Request to retrieve sequence rows for a single sequence.
 pub struct RetrieveSequenceRows {
+    /// Optional lower bound on row numbers.
     pub start: Option<i64>,
+    /// Optional upper bound on row numbers.
     pub end: Option<i64>,
+    /// Maximum number of rows to return per request.
     pub limit: Option<i32>,
+    /// Optional cursor for pagination.
     pub cursor: Option<String>,
+    /// Columns to retrieve values from.
     pub columns: Option<Vec<String>>,
     #[serde(flatten)]
+    /// ID of sequence to retrieve from.
     pub id: Identity,
 }
 
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Request to retrieve the sequence row with the highest row number.
 pub struct RetrieveLastSequenceRow {
+    /// Columns to retrieve from.
     pub columns: Option<Vec<String>>,
+    /// Retrieve the row with the highest row number that is smaller than this.
     pub before: Option<i64>,
     #[serde(flatten)]
+    /// ID of sequence to retrieve from.
     pub id: Identity,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
+/// Request to delete a list of sequence rows from a sequence.
 pub struct DeleteSequenceRows {
+    /// Row numbers to delete.
     pub rows: Vec<i64>,
     #[serde(flatten)]
+    /// ID of sequence to delete from.
     pub id: Identity,
 }
