+++ cognite/src/dto/data_modeling/mod.rs
@@ -4,5 +4,4 @@ pub mod data_models;
 pub mod instances;
 pub mod query;
 pub mod spaces;
-pub mod value;
 pub mod views;
