+++ cognite/src/dto/mod.rs
@@ -8,5 +8,6 @@ pub mod core;
 pub mod data_ingestion;
 pub mod data_modeling;
 pub mod data_organization;
+pub mod filter;
 pub mod iam;
 pub mod identity;
